+++ aim/web/ui/src/pages/Params/Params.tsx
@@ -61,6 +61,7 @@ const Params = ({
   resizeMode,
   notifyData,
   hiddenColumns,
+  hideSystemMetrics,
   onExportTableData,
   onCurveInterpolationChange,
   onActivePointChange,
@@ -234,6 +235,7 @@ const Params = ({
                   sortFields={sortFields}
                   hiddenRows={hiddenMetrics}
                   hiddenColumns={hiddenColumns}
+                  hideSystemMetrics={hideSystemMetrics}
                   resizeMode={resizeMode}
                   columnsWidths={columnsWidths}
                   selectedRows={selectedRows}