+++ aim/web/ui/src/services/models/explorer/createAppModel.ts
@@ -7,7 +7,7 @@ import { IPoint } from 'components/ScatterPlot';
 
 import COLORS from 'config/colors/colors';
 import DASH_ARRAYS from 'config/dash-arrays/dashArrays';
-import { ResizeModeEnum } from 'config/enums/tableEnums';
+import { HideColumnsEnum, ResizeModeEnum } from 'config/enums/tableEnums';
 import { AlignmentNotificationsEnum } from 'config/notification-messages/notificationMessages';
 import { RowHeightSize } from 'config/table/tableConfigs';
 import { DensityOptions } from 'config/enums/densityEnum';
@@ -340,6 +340,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
           config.table = {
             resizeMode: ResizeModeEnum.Resizable,
             rowHeight: RowHeightSize.md,
+            hideSystemMetrics: true,
             sortFields: [],
             hiddenMetrics: [],
             hiddenColumns: [],
@@ -939,7 +940,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
       const uniqContexts = _.uniq(contexts);
 
       const mappedData =
-        data.reduce((acc: any, item: any) => {
+        data?.reduce((acc: any, item: any) => {
           acc[item.hash] = { runHash: item.hash, ...item.props };
           return acc;
         }, {}) || {};
@@ -2072,7 +2073,6 @@ function createAppModel(appConfig: IAppInitialConfig) {
       default:
         return {};
     }
-
     function getRunsModelMethods() {
       let runsRequestRef: {
         call: (
@@ -2089,6 +2089,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
         abort: () => void;
       };
       let liveUpdateInstance: LiveUpdateService | null;
+      let updateTableTimeoutId: ReturnType<typeof setTimeout>;
 
       function initialize(appId: string = ''): {
         call: () => Promise<void>;
@@ -2218,7 +2219,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
                 model.getState()?.config?.table.columnsOrder!,
                 model.getState()?.config?.table.hiddenColumns!,
               );
-
+              updateTableData(tableData, tableColumns, configData);
               model.setState({
                 data,
                 selectedRows: shouldResetSelectedRows
@@ -2239,14 +2240,6 @@ function createAppModel(appConfig: IAppInitialConfig) {
                   },
                 },
               });
-              setTimeout(() => {
-                const tableRef: any = model.getState()?.refs?.tableRef;
-                tableRef.current?.updateData({
-                  newData: tableData.rows,
-                  newColumns: tableColumns,
-                  hiddenColumns: configData.table.hiddenColumns!,
-                });
-              }, 0);
             } catch (ex: Error | any) {
               if (ex.name === 'AbortError') {
                 // Abort Error
@@ -2278,12 +2271,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
           configData?.table?.columnsOrder!,
           configData?.table?.hiddenColumns!,
         );
-        const tableRef: any = model.getState()?.refs?.tableRef;
-        tableRef.current?.updateData({
-          newData: tableData.rows,
-          newColumns: tableColumns,
-          hiddenColumns: configData?.table?.hiddenColumns!,
-        });
+        updateTableData(tableData, tableColumns, configData);
         model.setState({
           config: configData,
           data,
@@ -2292,6 +2280,28 @@ function createAppModel(appConfig: IAppInitialConfig) {
           sameValueColumns: tableData.sameValueColumns,
           selectedRows,
         });
+      }
+
+      function updateTableData(
+        tableData: {
+          rows: any;
+          sameValueColumns: string[];
+        },
+        tableColumns: ITableColumn[],
+        configData: IAppModelConfig | any,
+      ): void {
+        if (updateTableTimeoutId) {
+          clearTimeout(updateTableTimeoutId);
+        }
+
+        updateTableTimeoutId = setTimeout(() => {
+          const tableRef: any = model.getState()?.refs?.tableRef;
+          tableRef.current?.updateData({
+            newData: tableData.rows,
+            newColumns: tableColumns,
+            hiddenColumns: configData.table.hiddenColumns!,
+          });
+        }, 0);
       }
 
       function prepareModelStateToCall(isInitial: boolean): IRunsAppModelState {
@@ -2354,7 +2364,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
         const processedData = groupData(runs);
         const uniqParams = _.uniq(params);
         const mappedData =
-          data.reduce((acc: any, item: any) => {
+          data?.reduce((acc: any, item: any) => {
             acc[item.hash] = { runHash: item.hash, ...item.props };
             return acc;
           }, {}) || {};
@@ -2943,6 +2953,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
         initialize,
         getRunsData,
         abortRequest,
+        updateModelData,
         getLastRunsData,
         onExportTableData,
         onNotificationDelete: onModelNotificationDelete,